requirements.txt
django-allauth=0.12.0